+++ tests/functional/test_assertion_failures.py
@@ -15,7 +15,8 @@ situations."""
 
 import pytest
 
-from mockify import Call, exc, satisfied
+from mockify import exc
+from mockify.core import Call, satisfied
 from mockify.actions import Return
 from mockify.cardinality import AtLeast, Exactly
 from mockify.mock import Mock