+++ tests/unit/core/test_session.py
@@ -10,7 +10,8 @@
 # ---------------------------------------------------------------------------
 import pytest
 
-from mockify import Call, Session, exc
+from mockify import exc
+from mockify.core import Call, Session
 from mockify.actions import Return
 from mockify.cardinality import Exactly