+++ tests/unit/mock/test_abc_mock.py
@@ -12,7 +12,8 @@ import abc
 
 import pytest
 
-from mockify import assert_satisfied, exc, satisfied
+from mockify import exc
+from mockify.core import assert_satisfied, satisfied
 from mockify.actions import Return
 from mockify.mock import ABCMock