+++ tests/unit/test_cardinality.py
@@ -11,7 +11,8 @@
 
 import pytest
 
-from mockify import exc, satisfied
+from mockify import exc
+from mockify.core import satisfied
 from mockify.actions import Return
 from mockify.cardinality import (
     ActualCallCount, AtLeast, AtMost, Between, Exactly